feat(py): allow passing binary body to the `data` parameter (#103)
Following the discussion under #97 , this PR widens the type accepted by
the `data` parameter.

`data` now accepts both a `dict[str,str]` and a binary representation of
the request body. This PR intentionally doesn't update the recently
added typings in the `impit.pyi` file, as this behaviour (accepting the
binary body) is considered deprecated in the `httpx` library we use as
the design master.